Update: Move worker-related code to own folder
